+++ src/main/java/com/zerobase/travel/festival_review/controller/FestivalReviewController.java
@@ -5,9 +5,11 @@ import com.zerobase.travel.festival_review.dto.FestivalReviewAddDTO;
 import com.zerobase.travel.festival_review.dto.FestivalReviewDTO;
 import com.zerobase.travel.festival_review.dto.FestivalReviewModifyDTO;
 import com.zerobase.travel.festival_review.service.FestivalReviewService;
+import com.zerobase.travel.user.dto.UserDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -37,13 +39,15 @@ public class FestivalReviewController {
     }
 
     @PostMapping("/festival/{festivalId}/review") // 축제 리뷰글 등록
-    public ResponseEntity<String> reviewInsert(@RequestBody FestivalReviewAddDTO reviewDTO, @PathVariable long festivalId) {
+    public ResponseEntity<String> reviewInsert(@RequestBody FestivalReviewAddDTO reviewDTO,
+                                               @PathVariable long festivalId,
+                                               @AuthenticationPrincipal UserDTO user) {
         FestivalReviewDTO festivalReviewDTO = FestivalReviewDTO.builder()
                 .reviewId(reviewDTO.getReviewId())
                 .festivalId(festivalId)
-                .userId(reviewDTO.getUserId())
+                .userId(user.getUserId())
                 .content(reviewDTO.getContent())
-                .writer(reviewDTO.getWriter())
+                .writer(user.getNickname())
                 .grade(reviewDTO.getGrade())
                 .createDate(LocalDate.now())
                 .build();
@@ -63,28 +67,35 @@ public class FestivalReviewController {
     @PutMapping("/festival/{festivalId}/review/{reviewId}") // 리뷰글 수정
     public ResponseEntity<String> reviewModify(@RequestBody FestivalReviewModifyDTO dto,
                                                @PathVariable long festivalId,
-                                               @PathVariable long reviewId) {
-        FestivalReviewDTO festivalReviewDTO = FestivalReviewDTO.builder()
-                .reviewId(reviewId)
-                .festivalId(festivalId)
-                .content(dto.getContent())
-                .grade(dto.getGrade())
-                .build();
+                                               @PathVariable long reviewId,
+                                               @AuthenticationPrincipal UserDTO user) {
+        if (service.get(reviewId).getUserId() == user.getUserId()) {
+            FestivalReviewDTO festivalReviewDTO = FestivalReviewDTO.builder()
+                    .reviewId(reviewId)
+                    .festivalId(festivalId)
+                    .content(dto.getContent())
+                    .grade(dto.getGrade())
+                    .build();
 
-        if(service.modify(festivalReviewDTO)) {
-            return ResponseEntity.status(HttpStatus.CREATED).body("리뷰글 수정 성공");
-        } else {
-            return ResponseEntity.status(HttpStatus.CREATED).body("리뷰글 수정 실패");
+            if (service.modify(festivalReviewDTO)) {
+                return ResponseEntity.status(HttpStatus.CREATED).body("리뷰글 수정 성공");
+            } else {
+                return ResponseEntity.status(HttpStatus.CREATED).body("리뷰글 수정 실패");
+            }
         }
+        return ResponseEntity.status(HttpStatus.CREATED).body("동일한 아이디가 아닙니다");
     }
     @DeleteMapping("/festival/{festivalId}/review/{reviewId}") // 리뷰글 삭제
     public ResponseEntity<String> reviewDelete(@PathVariable long festivalId,
-                                               @PathVariable long reviewId) {
-
-        if(service.delete(festivalId,reviewId)) {
-            return ResponseEntity.status(HttpStatus.CREATED).body("리뷰글 삭제 성공");
-        } else {
-            return ResponseEntity.status(HttpStatus.CREATED).body("이미 삭제되거나 없는 글입니다.");
+                                               @PathVariable long reviewId,
+                                               @AuthenticationPrincipal UserDTO user) {
+        if (service.get(reviewId).getUserId() == user.getUserId()) {
+            if (service.delete(festivalId, reviewId)) {
+                return ResponseEntity.status(HttpStatus.CREATED).body("리뷰글 삭제 성공");
+            } else {
+                return ResponseEntity.status(HttpStatus.CREATED).body("이미 삭제되거나 없는 글입니다.");
+            }
         }
+        return ResponseEntity.status(HttpStatus.CREATED).body("동일한 아이디가 아닙니다.");
     }
 }
